First major additions to resumé website
Created a home page with various images of different coding laguages and resources. Main implementation was Flex, use both rows and columns. Used flex columns to put the words underneath the pictures. I made it so when the screen is resized we don't lose any of the quote or the picture. Made it so when you hover over "Home" "Resume" Contact" "GitHub" it will bold the color. Made it so hyperlink text stays white before and after clicking. Made the websites icon my face.